requirements.txt
flask==3.0.0
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
replicate==0.22.0 